% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_split_ratio.R
\name{eval_split_ratio}
\alias{eval_split_ratio}
\title{Evaluate the performance of different training test split ratios on linear regression model}
\usage{
eval_split_ratio(x, y)
}
\arguments{
\item{x}{series of numerical values, independent variable in the linear regression model}

\item{y}{series of numerical values, the target variable in the prediction model}
}
\value{
the resulting box-plot of relationship between share of training data and the mean squared error of prediction
}
\description{
Evaluate the performance of different training test split ratios on linear regression model
}
\examples{
eval_split_ratio(Boston$dis, Boston$nox)

}
